Update SearchBar.js
Ln 82 flagged a Warning, which opened up varied approaches to resolving within React formats. Found it works fine to add the recommended comment just as is to the end of the line preceding the <a> tag, but with (/* */} JSX-format enclosures. Other JSX notation or target not required for successful compilation without warnings. *Issue apart: Favicon url may need resolving.